Removing unnecessary cast in hsv.cpp (#530)
Both temp_red and max_color are float32_t so the result of the subraction already has the right type.
This should also remove the only dependency on Boost.NumericConversion

+++ include/boost/gil/extension/toolbox/color_spaces/hsv.hpp
@@ -8,8 +8,6 @@
 #ifndef BOOST_GIL_EXTENSION_TOOLBOX_COLOR_SPACES_HSV_HPP
 #define BOOST_GIL_EXTENSION_TOOLBOX_COLOR_SPACES_HSV_HPP
 
-#include <boost/numeric/conversion/cast.hpp>
-
 #include <boost/gil/color_convert.hpp>
 #include <boost/gil/detail/mp11.hpp>
 #include <boost/gil/typedefs.hpp>
@@ -76,8 +74,7 @@ template <> struct default_color_converter_impl<rgb_t, hsv_t> {
       // it doesn't matter what value it has
       hue = 0.f;
     } else {
-      if ((std::abs)(boost::numeric_cast<float32_t>(temp_red - max_color)) <
-          0.0001f) {
+      if ((std::abs)(temp_red - max_color) < 0.0001f) {
         hue = (temp_green - temp_blue) / diff;
       } else if (temp_green >=
                  max_color) // means == but >= avoids compiler warning; color is
